Main thread is now SDL input thread.
This has been done because the Windows version of SDL
requires that the inputs that the user does over the application
should be done on the main thread. Since there was no reason
for the emulation thread to be the main one, I switched them
to be the SDL input the main thread, while the emulation thread
is the "Worker thread" as VC++ stands.
Signed-off-by: Rodrigo Javier Tobar Carrizo <rtobar@csrg.inf.utfsm.cl>